# frontend/app/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoLD API",
    "description": "\n    # CoLD API Documentation\n    The Choice of Law Dataverse (CoLD) API provides programmatic access to a curated knowledge base on choice of law in international contracts.\n\n    Core capabilities:\n    - Full-text search across all data domains with filtering and sorting by date.\n    - Fetch curated details for a specific record by CoLD ID, including first-hop related entries.\n    - Retrieve full tables or filtered subsets using user-facing fields (mapping-aware).\n    - Site map and landing page helper endpoints for the frontend.\n\n    Authentication:\n    - All endpoints (except the root) require a Bearer JWT in the `Authorization` header: `Authorization: Bearer <token>`.\n    ",
    "contact": {
      "name": "CoLD Team",
      "url": "https://choice-of-law-dataverse.org/"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/search/": {
      "post": {
        "tags": ["Search"],
        "summary": "Full-text search across CoLD data",
        "description": "Searches across multiple domains (Answers, HCCH Answers, Court Decisions, Domestic Instruments, Regional Instruments, International Instruments, and Literature). Filters support user-facing fields like 'tables', 'Jurisdictions', or 'Themes'. You can sort by date and paginate results.",
        "operationId": "handle_full_text_search_api_v1_search__post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FullTextSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results including pagination and total matches.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullTextSearchResponse"
                },
                "example": {
                  "test": false,
                  "total_matches": 2,
                  "page": 1,
                  "page_size": 2,
                  "results": [
                    {
                      "source_table": "Answers",
                      "id": "CHE_15-TC",
                      "Title": "\u2026"
                    },
                    {
                      "source_table": "Court Decisions",
                      "id": "CD-GBR-1167",
                      "Title": "\u2026"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search/details": {
      "post": {
        "tags": ["Search"],
        "summary": "Fetch a curated record by CoLD ID including hop-1 relations",
        "description": "Given a user-facing table and a CoLD ID, returns the canonical record and its first-hop related entries.",
        "operationId": "handle_curated_details_search_api_v1_search_details_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CuratedDetailsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Flattened, mapping-transformed record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuratedDetailsRecord"
                },
                "example": {
                  "source_table": "Answers",
                  "id": "CHE_15-TC",
                  "Title": "\u2026",
                  "hop1_relations": {}
                }
              }
            }
          },
          "404": {
            "description": "Record not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search/full_table": {
      "post": {
        "tags": ["Search"],
        "summary": "Return full or filtered table",
        "description": "Returns all records from the specified table or a filtered subset. Filters accept user-facing field names and values (mapping-aware).",
        "operationId": "return_full_table_api_v1_search_full_table_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FullTableRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Array of transformed records from the requested table.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/AnswerRecord"
                      },
                      {
                        "$ref": "#/components/schemas/CourtDecisionRecord"
                      },
                      {
                        "$ref": "#/components/schemas/DomesticInstrumentRecord"
                      },
                      {
                        "$ref": "#/components/schemas/InternationalInstrumentRecord"
                      },
                      {
                        "$ref": "#/components/schemas/RegionalInstrumentRecord"
                      },
                      {
                        "$ref": "#/components/schemas/LiteratureRecord"
                      },
                      {
                        "$ref": "#/components/schemas/ArbitralAwardRecord"
                      },
                      {
                        "$ref": "#/components/schemas/ArbitralInstitutionRecord"
                      },
                      {
                        "$ref": "#/components/schemas/ArbitralProvisionRecord"
                      },
                      {
                        "$ref": "#/components/schemas/ArbitralRuleRecord"
                      },
                      {
                        "$ref": "#/components/schemas/DomesticLegalProvisionRecord"
                      },
                      {
                        "$ref": "#/components/schemas/InternationalLegalProvisionRecord"
                      },
                      {
                        "$ref": "#/components/schemas/RegionalLegalProvisionRecord"
                      },
                      {
                        "$ref": "#/components/schemas/JurisdictionRecord"
                      },
                      {
                        "$ref": "#/components/schemas/QuestionRecord"
                      },
                      {
                        "$ref": "#/components/schemas/RecordBase"
                      }
                    ]
                  },
                  "title": "Response Return Full Table Api V1 Search Full Table Post"
                },
                "example": [
                  {
                    "source_table": "Answers",
                    "id": "CHE_15-TC",
                    "Jurisdictions": ["Switzerland"],
                    "Title": "\u2026"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Missing or invalid table parameter."
          },
          "500": {
            "description": "Server error while querying the table."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search/specialists/{jurisdiction_alpha_code}": {
      "get": {
        "tags": ["Search"],
        "summary": "Get specialists by jurisdiction",
        "description": "Returns all specialists associated with a specific jurisdiction using Alpha_3_Code.",
        "operationId": "get_specialists_by_jurisdiction_api_v1_search_specialists__jurisdiction_alpha_code__get",
        "parameters": [
          {
            "name": "jurisdiction_alpha_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Jurisdiction Alpha Code"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of specialists for the given jurisdiction.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecialistResponse"
                  },
                  "title": "Response Get Specialists By Jurisdiction Api V1 Search Specialists  Jurisdiction Alpha Code  Get"
                },
                "example": [
                  {
                    "id": 1,
                    "created_at": "2024-01-01T00:00:00",
                    "updated_at": "2024-01-15T10:30:00",
                    "created_by": "admin",
                    "updated_by": "admin",
                    "nc_order": 1.0,
                    "ncRecordId": "rec123",
                    "ncRecordHash": "hash123",
                    "Specialist": "Dr. Jane Smith",
                    "Created": "2024-01-01T00:00:00"
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Jurisdiction not found."
          },
          "500": {
            "description": "Server error while querying specialists."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CuratedDetailsRecord": {
        "properties": {
          "sourceTable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcetable"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "rank": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "coldId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coldid"
          },
          "hop1Relations": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "type": "object"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hop1Relations"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "CuratedDetailsRecord"
      },
      "CuratedDetailsRequest": {
        "properties": {
          "table": {
            "type": "string",
            "title": "Table"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "response_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["parsed", "raw", "both"]
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Type",
            "description": "Select 'parsed' (default), 'raw', or 'both' for response data.",
            "default": "parsed"
          }
        },
        "type": "object",
        "required": ["table", "id"],
        "title": "CuratedDetailsRequest",
        "examples": [
          {
            "id": "CHE_15-TC",
            "response_type": "parsed",
            "table": "Answers"
          }
        ]
      },
...
      "FullTableRequest": {
        "properties": {
          "table": {
            "type": "string",
            "title": "Table"
          },
          "filters": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FTFilterOption"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          },
          "response_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["parsed", "raw", "both"]
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Type",
            "description": "Select 'parsed' (default), 'raw', or 'both' for response data.",
            "default": "parsed"
          }
        },
        "type": "object",
        "required": ["table"],
        "title": "FullTableRequest",
        "examples": [
          {
            "filters": [
              {
                "column": "Jurisdictions",
                "value": "Switzerland"
              }
            ],
            "response_type": "parsed",
            "table": "Answers"
          }
        ]
      },
      "FullTextSearchRequest": {
        "properties": {
          "search_string": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search String"
          },
          "filters": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FTSFilterOption"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "Page number, must be >= 1",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "description": "Number of results per page",
            "default": 50
          },
          "sort_by_date": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort By Date",
            "description": "Sort results by date descending if True.",
            "default": false
          },
          "response_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["parsed", "raw", "both"]
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Type",
            "description": "Select 'parsed' (default), 'raw', or 'both' for response data.",
            "default": "parsed"
          }
        },
        "type": "object",
        "title": "FullTextSearchRequest",
        "examples": [
          {
            "filters": [
              {
                "column": "tables",
                "values": ["Answers", "Court Decisions"]
              },
              {
                "column": "jurisdictions",
                "values": ["Switzerland", "France"]
              },
              {
                "column": "Themes",
                "values": ["Party autonomy", "Freedom of choice"]
              }
            ],
            "page": 1,
            "page_size": 100,
            "response_type": "parsed",
            "search_string": "",
            "sort_by_date": true
          }
        ]
      },
      "FullTextSearchResponse": {
        "properties": {
          "test": {
            "type": "boolean",
            "title": "Test"
          },
          "totalMatches": {
            "type": "integer",
            "title": "Totalmatches"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          },
          "results": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AnswerRecord"
                },
                {
                  "$ref": "#/components/schemas/CourtDecisionRecord"
                },
                {
                  "$ref": "#/components/schemas/DomesticInstrumentRecord"
                },
                {
                  "$ref": "#/components/schemas/InternationalInstrumentRecord"
                },
                {
                  "$ref": "#/components/schemas/RegionalInstrumentRecord"
                },
                {
                  "$ref": "#/components/schemas/LiteratureRecord"
                },
                {
                  "$ref": "#/components/schemas/ArbitralAwardRecord"
                },
                {
                  "$ref": "#/components/schemas/ArbitralInstitutionRecord"
                },
                {
                  "$ref": "#/components/schemas/ArbitralProvisionRecord"
                },
                {
                  "$ref": "#/components/schemas/ArbitralRuleRecord"
                },
                {
                  "$ref": "#/components/schemas/DomesticLegalProvisionRecord"
                },
                {
                  "$ref": "#/components/schemas/InternationalLegalProvisionRecord"
                },
                {
                  "$ref": "#/components/schemas/RegionalLegalProvisionRecord"
                },
                {
                  "$ref": "#/components/schemas/JurisdictionRecord"
                },
                {
                  "$ref": "#/components/schemas/QuestionRecord"
                },
                {
                  "$ref": "#/components/schemas/RecordBase"
                }
              ]
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": ["test", "totalMatches", "page", "pageSize", "results"],
        "title": "FullTextSearchResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecordBase": {
        "properties": {
          "sourceTable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcetable"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "rank": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "RecordBase"
      },
      "RegionalInstrumentRecord": {
        "properties": {
          "sourceTable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcetable"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "rank": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "idNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idnumber"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "abbreviation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abbreviation"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "attachment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachment"
          },
          "recordId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recordid"
          },
          "created": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created"
          },
          "lastModified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastmodified"
          },
          "lastModifiedById": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastmodifiedbyid"
          },
          "createdById": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdbyid"
          },
          "sortDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sortdate"
          },
          "specialists": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialists"
          },
          "specialistsLink": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialistslink"
          },
          "regionalLegalProvisions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regionallegalprovisions"
          },
          "regionalLegalProvisionsLink": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regionallegalprovisionslink"
          },
          "lastModifiedByEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastmodifiedbyemail"
          },
          "lastModifiedByName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastmodifiedbyname"
          },
          "createdByEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdbyemail"
          },
          "createdByName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdbyname"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "RegionalInstrumentRecord"
      },
...
    }
  },
...
  "servers": [
    {
      "url": "/api/proxy"
    }
  ]
}